project/pyproject.toml
[tool.poetry]
name = "leetcode-bot"
version = "0.1.0"
description = ""
authors = ["FawenYo <twsky020600@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
line-bot-sdk = "^1.19.0"
APScheduler = "^3.7.0"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
pymongo = {extras = ["srv"], version = "^3.11.3"}
python-dotenv = "^0.15.0"
rich = "^9.13.0"
aiofiles = "^0.6.0"
Jinja2 = "^2.11.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
